refactor: move SizeBadge to shared location and add rank-based sorting
- Move SizeBadge from reviews/+components to $orgSlug/+components
- Use cn() for className composition
- Remove unnecessary hover styles on non-interactive badges
- Add sortingFn for Size column (XS→S→M→L→XL→unclassified)
