+++ app/routes/$orgSlug/_index/+columns.tsx
@@ -3,7 +3,7 @@ import { AppSortableHeader } from '~/app/components'
 import { Badge, HStack } from '~/app/components/ui'
 import { Avatar, AvatarFallback, AvatarImage } from '~/app/components/ui/avatar'
 import dayjs from '~/app/libs/dayjs'
-import { SizeBadge } from '../reviews/+components/size-badge'
+import { SizeBadge } from '../+components/size-badge'
 import type { PullRequest } from './index'
 
 export const columns: ColumnDef<PullRequest>[] = [
@@ -62,6 +62,12 @@ export const columns: ColumnDef<PullRequest>[] = [
     accessorKey: 'complexity',
     header: ({ column }) => <AppSortableHeader column={column} title="Size" />,
     cell: ({ row }) => <SizeBadge complexity={row.original.complexity} />,
+    sortingFn: (a, b) => {
+      const rank: Record<string, number> = { XS: 0, S: 1, M: 2, L: 3, XL: 4 }
+      const aRank = rank[a.original.complexity ?? ''] ?? 5
+      const bRank = rank[b.original.complexity ?? ''] ?? 5
+      return aRank - bRank
+    },
     enableHiding: false,
   },
   {
