+++ app/routes/$orgSlug/ongoing/+columns.tsx
@@ -2,7 +2,7 @@ import type { ColumnDef } from '@tanstack/react-table'
 import { AppSortableHeader } from '~/app/components'
 import { Avatar, AvatarFallback, AvatarImage } from '~/app/components/ui/avatar'
 import dayjs from '~/app/libs/dayjs'
-import { SizeBadge } from '../reviews/+components/size-badge'
+import { SizeBadge } from '../+components/size-badge'
 import type { PullRequest } from './index'
 
 export const columns: ColumnDef<PullRequest>[] = [
@@ -61,6 +61,12 @@ export const columns: ColumnDef<PullRequest>[] = [
     accessorKey: 'complexity',
     header: ({ column }) => <AppSortableHeader column={column} title="Size" />,
     cell: ({ row }) => <SizeBadge complexity={row.original.complexity} />,
+    sortingFn: (a, b) => {
+      const rank: Record<string, number> = { XS: 0, S: 1, M: 2, L: 3, XL: 4 }
+      const aRank = rank[a.original.complexity ?? ''] ?? 5
+      const bRank = rank[b.original.complexity ?? ''] ?? 5
+      return aRank - bRank
+    },
     enableHiding: false,
   },
   {
